server/requirements.txt
Flask==2.3.3
langchain==0.0.273
llama-cpp-python==0.1.78
mmengine==0.8.4
mmcv==2.0.1
-e ./mmsegmentation/
opencv-python==4.8.0.76
python-dotenv==1.0.0
torch==2.0.1
torchvision==0.15.2
